fix: removes calendar also from SelectedCalendar once deleted

+++ packages/trpc/server/routers/loggedInViewer/deleteCredential.handler.ts
@@ -104,40 +104,21 @@ export const deleteCredentialHandler = async ({ ctx, input }: DeleteCredentialOp
     }
 
     // If it's a calendar, remove the destination calendar from the event type
-    if (credential.app?.categories.includes(AppCategories.calendar)) {
-      if (eventType.destinationCalendar?.credential?.appId === credential.appId) {
-        const destinationCalendar = await prisma.destinationCalendar.findFirst({
+    if (
+      credential.app?.categories.includes(AppCategories.calendar) &&
+      eventType.destinationCalendar?.credential?.appId === credential.appId
+    ) {
+      const destinationCalendar = await prisma.destinationCalendar.findFirst({
+        where: {
+          id: eventType.destinationCalendar?.id,
+        },
+      });
+      if (destinationCalendar) {
+        await prisma.destinationCalendar.delete({
           where: {
-            id: eventType.destinationCalendar?.id,
+            id: destinationCalendar.id,
           },
         });
-        if (destinationCalendar) {
-          await prisma.destinationCalendar.delete({
-            where: {
-              id: destinationCalendar.id,
-            },
-          });
-        }
-      }
-
-      if (externalId) {
-        const existingSelectedCalendar = await prisma.selectedCalendar.findFirst({
-          where: {
-            externalId: externalId,
-          },
-        });
-        // @TODO: SelectedCalendar doesn't have unique ID so we should only delete one item
-        if (existingSelectedCalendar) {
-          await prisma.selectedCalendar.delete({
-            where: {
-              userId_integration_externalId: {
-                userId: existingSelectedCalendar.userId,
-                externalId: existingSelectedCalendar.externalId,
-                integration: existingSelectedCalendar.integration,
-              },
-            },
-          });
-        }
       }
     }
 
@@ -307,6 +288,27 @@ export const deleteCredentialHandler = async ({ ctx, input }: DeleteCredentialOp
           }
         });
       }
+    }
+  }
+
+  // If it's a calendar remove it from the SelectedCalendars
+  if (credential.app?.categories.includes(AppCategories.calendar) && externalId) {
+    const existingSelectedCalendar = await prisma.selectedCalendar.findFirst({
+      where: {
+        externalId,
+      },
+    });
+
+    if (existingSelectedCalendar) {
+      await prisma.selectedCalendar.delete({
+        where: {
+          userId_integration_externalId: {
+            userId: existingSelectedCalendar.userId,
+            externalId: existingSelectedCalendar.externalId,
+            integration: existingSelectedCalendar.integration,
+          },
+        },
+      });
     }
   }
